ERPHI-67: 15m 70pmd se mejora método para complementar datos en el modelo CFDSAT para eliminar conceptos y traslados únicamente si los traslados estan cargados sin datos en impuesto ocasionado por obtención de datos incompleta de los CFDI cuando se cargan procesando directorio.

+++ app/Models/SEGURIDAD_ERP/Contabilidad/CFDSAT.php
@@ -66,7 +66,7 @@ class CFDSAT extends Model
     ];
 
     protected $dates =["fecha", "fecha_cancelacion"];
-    protected $dateFormat = 'Y-m-d H:i:s';
+    //protected $dateFormat = 'Y-m-d H:i:s';
 
     public function carga()
     {
@@ -314,23 +314,31 @@ class CFDSAT extends Model
                 $this->save();
             }
 
-            if(key_exists("conceptos",$data)){
-                $this->conceptos()->delete();
-                foreach($data["conceptos"] as $concepto){
-                    $conceptoObj = $this->conceptos()->create($concepto);
-                    if(key_exists("traslados",$concepto)){
-                        foreach($concepto["traslados"] as $traslado){
-                            $conceptoObj->traslados()->create($traslado);
+            $concepto = $this->conceptos()->first();
+            $traslado = $concepto->traslados()->first();
+
+            if($traslado){
+                if($traslado->impuesto === null){
+                    if(key_exists("conceptos",$data)){
+                        $this->conceptos()->delete();
+                        foreach($data["conceptos"] as $concepto){
+                            $conceptoObj = $this->conceptos()->create($concepto);
+                            if(key_exists("traslados",$concepto)){
+                                foreach($concepto["traslados"] as $traslado){
+                                    $conceptoObj->traslados()->create($traslado);
+                                }
+                            }
+                        }
+                    }
+                    if(key_exists("traslados",$data)){
+                        $this->traslados()->delete();
+                        foreach($data["traslados"] as $traslado){
+                            $this->traslados()->create($traslado);
                         }
                     }
                 }
             }
-            if(key_exists("traslados",$data)){
-                $this->traslados()->delete();
-                foreach($data["traslados"] as $traslado){
-                    $this->traslados()->create($traslado);
-                }
-            }
+
             if(key_exists("documentos_pagados",$data)){
                 $this->documentosPagados()->delete();
                 foreach($data["documentos_pagados"] as $documento_pagado){
